Move the segment fetch thunks into the reference module
